Move SDK version logic from device.go
